Move profile extraction code to lib
